requirements.txt
pymongo==4.6.1
dnspython==2.4.2
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
jupyter==1.0.0
python-dotenv==1.0.0
requests==2.31.0
